♻️ refactor: 공용 타입 components/report/types → utils/types/meal.ts로 이동
